requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via django
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
build==1.2.2.post1
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via pip-tools
coverage[toml]==7.6.10
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
django==4.2.17
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-cors-headers
    #   django-dramatiq
    #   django-filter
    #   django-waffle
    #   djangorestframework
    #   drf-spectacular
django-appconf==1.0.6
    # via django-imagekit
django-cors-headers==4.3.0
    # via -r requirements.in
django-dramatiq==0.12.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-imagekit==5.0.0
    # via -r requirements.in
django-waffle==4.1.0
    # via -r requirements.in
djangorestframework==3.15.0
    # via
    #   -r requirements.in
    #   drf-spectacular
dramatiq[rabbitmq]==1.16.0
    # via
    #   -r requirements.in
    #   django-dramatiq
drf-spectacular==0.28.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via pytest
filelock==3.17.0
    # via virtualenv
gevent==22.10.2
    # via -r requirements.in
greenlet==3.1.1
    # via gevent
gunicorn==21.2.0
    # via -r requirements.in
identify==2.6.6
    # via pre-commit
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
jsonschema==4.22.0
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   build
    #   gunicorn
    #   pytest
pika==1.3.2
    # via dramatiq
pilkit==3.0
    # via django-imagekit
pillow==11.1.0
    # via pilkit
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==2.19.0
    # via -r requirements.in
prometheus-client==0.21.1
    # via dramatiq
psycopg2==2.9.9
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.2.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-cov==5.0.0
    # via -r requirements.in
pytest-django==4.8.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   pre-commit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruff==0.7.2
    # via -r requirements.in
sqlparse==0.5.3
    # via django
toml==0.10.2
    # via pre-commit
tomli==2.2.1
    # via
    #   build
    #   coverage
    #   pip-tools
    #   pytest
typing-extensions==4.12.2
    # via
    #   asgiref
    #   referencing
uritemplate==4.1.1
    # via drf-spectacular
virtualenv==20.29.1
    # via pre-commit
wheel==0.45.1
    # via pip-tools
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
